Final stabilization and optimization of ESP32 synthesizer
- Optimized GranularExciter with lookup tables and fast PRNG; fixed compilation error.
- Fixed signed/unsigned casting in FM synthesis for both Voice and FMPair modules.
- Implemented soft clipping (tanhf) at the master output for improved sound quality.
- Increased stack sizes for audio (20KB) and display (8KB) tasks to prevent crashes.
- Increased DMA buffer size to 128 samples to reduce CPU overhead.
- Expanded internal test suite to include Sync and Ring Modulation tests.
- Fixed and separated serial commands for Granular, Resonator, and Glide.
- Implemented melodic tracking for the global Resonator.

+++ Synthesis.cpp
@@ -118,14 +118,14 @@ void FMPair::setFrequency(float hz) {
 
 float FMPair::process() {
     // Modulator with feedback
-    uint32_t fbOffset = (uint32_t)(lastModOut_ * feedback_ * modInc_ * 2.0f);
-    float modOut = Wavetables::readSine(modPhase_ + fbOffset);
+    int32_t fbOffset = (int32_t)(lastModOut_ * feedback_ * (float)modInc_ * 2.0f);
+    float modOut = Wavetables::readSine((uint32_t)((int32_t)modPhase_ + fbOffset));
     lastModOut_ = modOut;
     modPhase_ += modInc_;
     
     // FM: modulator affects carrier phase
-    uint32_t fmOffset = (uint32_t)(modOut * modIndex_ * carrierInc_);
-    float carrierOut = Wavetables::readSine(carrierPhase_ + fmOffset);
+    int32_t fmOffset = (int32_t)(modOut * modIndex_ * (float)carrierInc_);
+    float carrierOut = Wavetables::readSine((uint32_t)((int32_t)carrierPhase_ + fmOffset));
     carrierPhase_ += carrierInc_;
     
     return carrierOut;
